src/queue-db/tests/requirements.txt
# Copyright 2024 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# Testing dependencies for queue-db microservice

# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Database testing
psycopg2-binary>=2.9.0
pytest-postgresql>=4.1.0
testcontainers>=3.4.0

# Performance testing
pytest-benchmark>=4.0.0
pytest-xdist>=3.0.0  # For parallel test execution

# Code quality
pytest-flake8>=1.1.0
pytest-mypy>=0.10.0
pytest-black>=0.3.0

# Test reporting
pytest-html>=3.1.0
pytest-json-report>=1.5.0

# Load testing
locust>=2.15.0

# Utilities
faker>=18.0.0  # For generating test data
freezegun>=1.2.0  # For time mocking
